Pick up WNBD and Ceph changes
WNBD now provides a DLL (libwnbd.dll) that handles all the driver
interaction. We'll make sure that we include the new files as well
as the symbols.

Instead of picking each symbol file individually, we'll just take
all the files that were copied to the "Symbols" dir. This way,
we allow adding new symbols (even Ceph symbols) without changing the
MSI definition. We're droping the separate .gitignore files to avoid
including them in the MSI.

We'll have to include the runtime as well. We assume that WNBD was
also built using Visual Studio 2019, so we include VC142.

rbd-nbd is now rbd-wnbd, so we need to update the ceph-rbd service.

While at it, we're updating the exit dialog text, as the ceph service
will work without requiring a valid ceph.conf.