Configurable Ceph build type
We're updating the Ceph build command to use a dynamic build by
default. Also, we're adding a couple of config options, specifying
whether we should include debug symbols and if so, whether to
include the minimum amount of debug information possible.

The "STRIP_ZIPPED" flag is dropped. If debug symbols are not needed,
it's better not to generate them at all.