Schedule a reboot after (un)installing the driver
The WNBD driver can fail to install if the host hasn't been
rebooted after a previous version was uninstalled. At the same time,
the adapter may be unavailable if a reboot is required after a
successful install.

To improve the user experience, we'll always ask for a reboot after
installing or uninstalling the driver.